Emma's Notebook (java8) [10/10] Accepted

+++ codestorm/emmas-notebook.java
@@ -5,7 +5,7 @@ public class Solution {
 
     public static void main(String[] args) {
         int t = new Scanner(System.in).nextInt();
-        int sum = 1;
+        long sum = 1;
         for (int i = 2, inc = 1; i <= t; ++i) {
             if (i % 2 == 0) {
                 ++inc;
